requirements.txt
Pillow==9.2.0
pydantic==1.9.2
